⤴️ Fixes to Connect 3.2
Debezium 1.9.5 switched to Kafka Connect 3.2.0. In that version, the `FileStream` connectors have been removed from the classpath, because of [KAFKA-13748](https://issues.apache.org/jira/browse/KAFKA-13748).

Kcctl integration tests automatically pick the latest Debezium so some of the tests that rely on `FileStreamSource` are not working anymore.

To address this issue, I decided to update the integration tests to instead use `MirrorHeartbeatConnector`.

Another alternative would be to update `debezium/connect` to optionally re-enable the `FileStream` connectors. Since this is only fixing tests and this has some security implications I decided to not follow this approach. If you think this would be preferable I can make the changes.