#353 Fix flaky integration tests
The DescribeConnectorCommand is manually invoked by the PatchConnectorCommand,
but without being given to a wrapping CommandLine instance. As a result, its
@CommandLine.Spec-annotated spec field is not initialized in that case, and
under some rare circumstances, the command will try to emit output via
spec.commandLine().getOut()::print (or println, or printf, etc.). This causes
a NullPointerException.

The scenario that leads to the attempted output is when the Connect cluster
serves a 409 response due to a pending rebalance. This is otherwise gracefully
handled with bounded retry logic by the PatchConnectorCommand.

This fix is a bit quick-and-dirty in that it doesn't prevent the
NullPointerException from occurring if the connector is not found, but that
situation should never arise except from extremely rare concurrent operations
on the Connect cluster. We can and should investigate a more fully-fledged
fix, but this is one quick and lightweight option that may suffice for now to
patch the recently-occurring integration test flakiness.